requirements.txt
django==3.2.3
django-phonenumber-field[phonenumbers]==5.2.0